requirements.txt
keyboard
psutil
obs-websocket-py
pywin32
Pillow